Changed 'lock' functionality
The lock functionality did not behave as I (or others) expected it to. Rather than suspend the session, which induces an animation and 'throws' the user back to the login-screen, I've implemented the same lock functionality that we can find in the Keychain menu-bar's 'Lock Screen' function. As the new method is only available to OS 10.9+, I've fallen back to the original function when it detects an older version of OS X.